requirements.txt
allure-pytest==2.13.2
pre-commit==3.3.3
pytest==7.1.2
pytest-check==2.1.5
requests==2.28.1
